+++ GridboxConnectorAddon-edge/GridboxConnector/main_test.py
@@ -120,5 +120,19 @@ class TestGridboxConnectorMethods(unittest.TestCase):
             viessmann_gridbox_connector.update_sensors(result[0])
             mock_heater_sensor.assert_called_once_with(3676,70.9)
 
+    def test_logger(self):
+        import os
+        import logging
+        from utils import SensitiveDataFilter
+        logger = logging.getLogger(__name__)
+        logger.setLevel(logging.getLevelName(os.getenv('LOG_LEVEL', 'INFO')))
+        formatter = logging.Formatter('%(asctime)s %(filename)s:%(lineno)d %(levelname)s - %(message)s')
+        console_handler = logging.StreamHandler()
+        console_handler.setFormatter(formatter)
+        logger.addHandler(console_handler)
+        logger.addFilter(SensitiveDataFilter())
+        login_message = "{'grant_type': 'http://auth0.com/oauth/grant-type/password-realm', 'username': 'username@username', 'password': 'UltraSecret', 'audience': 'my.gridx', 'client_id': 'oZpr934Ikn8OZOHTJEcrgXkjio0I0Q7b', 'scope': 'email openid', 'realm': 'viessmann-authentication-db', 'client_secret': ''}"
+        logger.info(login_message)
+
 if __name__ == '__main__':
     unittest.main()