+++ GridboxConnectorAddon-edge/GridboxConnector/utils.py
@@ -7,16 +7,17 @@ class SensitiveDataFilter(logging.Filter):
     def filter(self, record):
         message = record.getMessage()
         try:
-            message_json = json.loads(message)
+            # Versuche die Nachricht als Python-Dictionary zu parsen
+            message_dict = ast.literal_eval(message)
             # Sensible Daten filtern, falls vorhanden
             sensitive_keys = ['username', 'password', 'id_token', 'access_token', 'client_id']
             for key in sensitive_keys:
-                if key in message_json:
-                    message_json[key] = '***'
-            # Das modifizierte JSON-Objekt zurück in einen String konvertieren
-            record.msg = json.dumps(message_json)
-        except json.JSONDecodeError:
-            logging.error("Error decoding JSON message")
+                if key in message_dict:
+                    message_dict[key] = '***'
+            # Das modifizierte Dictionary zurück in einen String konvertieren
+            record.msg = str(message_dict)
+        except (ValueError, SyntaxError):
+            logging.error(f"Error parsing message: {message}")
         except Exception as e:
             logging.error(f"Error filtering sensitive data: {e}")
         return True